requirements.txt
Django==2.2.7
djangorestframework==3.11.0
pytz==2019.3
sqlparse==0.3.0
